* Wad manager window is now centered * Maps with over 32767 sidedefs no longer crash SLADE + Added some checks to prevent crashes if map data is corrupted + Added 'recent wads' list to wad manager window

+++ trunk/linedef.h
@@ -27,8 +27,8 @@ struct linedef_t
 	WORD	vertex2;
 	WORD	flags;
 	WORD	type;
-	short	side1;
-	short	side2;
+	long	side1;
+	long	side2;
 
 	// Doom format only
 	short	sector_tag;
